Optimize theme handling
