+++ .changeset/odd-lions-provide.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": patch
+---
+
+Optimize theme handling
