--- apps/nextjs-website/src/app/theme.ts
@@ -1,165 +1,0 @@
-'use client';
-import { createTheme } from '@mui/material/styles';
-import { theme as MUIItaliaTheme } from '@pagopa/mui-italia';
-import { FONT_THEME_EXPERIMENTAL } from '@react-components/components/common/Common.helpers';
-
-declare module '@mui/material/styles' {
-  interface Palette {
-    readonly custom: {
-      readonly backgroundColorDark: string;
-      readonly backgroundColorDarkAlternative: string;
-      readonly backgroundColorLight: string;
-      readonly backgroundColorLightGrey: string;
-      readonly divider: string;
-      readonly primaryColorDark: string;
-      readonly walletPrimaryColor: string;
-      readonly walletPrimaryColorHover: string;
-      readonly walletBackgroundColorDark: string;
-      readonly backgroundColorDefault: string;
-      readonly darkLinkColor: string;
-      readonly grayLinkColor: string;
-      readonly black50: string;
-      readonly editorialSwitchButtonsBackgroundLightBlue: string;
-      readonly editorialSwitchButtonsBackgroundWhite: string;
-      readonly bannerLinkLightBlueIO: string;
-      readonly bannerLinkDarkBlueIO: string;
-      readonly bannerLinkLightBlueSend: string;
-      readonly bannerLinkDarkBlueSend: string;
-      readonly bannerLinkLightBlueWallet: string;
-      readonly bannerLinkDarkBlueWallet: string;
-      readonly blueIO: {
-        readonly 500: string;
-      };
-      readonly blueWallet: {
-        readonly 500: string;
-      };
-      readonly highLightBoxLightGreenBackground: string;
-      readonly highLightBoxPurpleText: string;
-      readonly highLightBoxLightPurpleBackground: string;
-      readonly black: string;
-      readonly white: string;
-      readonly matteWhiteBorder: string;
-      readonly ioChipsBackgroundColorLight: string;
-      readonly walletChipsBackgroundColorLight: string;
-      readonly sendChipsBackgroundColorLight: string;
-      readonly chipsBackgroundColorDark: string;
-      readonly ioChipsBackgroundColorLightHover: string;
-      readonly walletChipsBackgroundColorLightHover: string;
-      readonly sendChipsBackgroundColorLightHover: string;
-      readonly chipsBackgroundColorDarkHover: string;
-      readonly chipsTextColor: string;
-      readonly borderColor: string;
-      readonly outlineColor: string;
-      readonly walletOutlineColor: string;
-      readonly ctaContainedLightHoverColor: string;
-      readonly ctaContainedDarkHoverColor: string;
-    };
-  }
-  // interface PaletteOptions {
-  //   readonly custom?: {
-  //     readonly backgroundColorDark?: string;
-  //     readonly backgroundColorLightGrey?: string;
-  //     readonly divider?: string;
-  //     readonly primaryColorDark?: string;
-  //     readonly backgroundColorDefault?: string;
-  //     readonly darkLinkColor?: string;
-  //     readonly grayLinkColor?: string;
-  //     readonly black50?: string;
-  //     readonly editorialSwitchButtonsBackgroundLightBlue?: string;
-  //     readonly editorialSwitchButtonsBackgroundWhite?: string;
-  //     readonly bannerLinkLightBlue?: string;
-  //     readonly bannerLinkDarkBlue?: string;
-  //   };
-  // }
-}
-
-const themeStyles = {
-  shadows: {
-    custom: {
-      boxShadow: '-6px -6px 19px 2px #002B551A',
-      otMenuMobile:
-        'rgb(0 43 85 / 10%) 0px 2px 4px -1px, rgb(0 43 85 / 5%) 0px 4px 5px',
-    },
-  },
-  palette: {
-    error: {
-      main: '#D13333',
-      dark: '#D13333',
-    },
-    custom: {
-      backgroundColorDark: '#0B3EE3',
-      backgroundColorDarkAlternative: '#0062C3',
-      backgroundColorLight: '#FFF',
-      backgroundColorLightGrey: '#FAFAFA',
-      divider: '#E3E7EB',
-      primaryColorDark: '#0B3EE3',
-      walletPrimaryColor: '#0066CC',
-      walletPrimaryColorHover: '#004D99',
-      walletBackgroundColorDark: '#0066CC',
-      backgroundColorDefault: '#FAFAFA',
-      darkLinkColor: '#1A73E8',
-      grayLinkColor: '#F4F5F8',
-      black50: '#00000050',
-      editorialSwitchButtonsBackgroundLightBlue: '#0073e61a',
-      editorialSwitchButtonsBackgroundWhite: '#ffffff1a',
-      bannerLinkLightBlueIO: '#1043e8',
-      bannerLinkDarkBlueIO: '#0B3EE3',
-      bannerLinkLightBlueSend: '#0A64D2',
-      bannerLinkDarkBlueSend: '#0062C3',
-      bannerLinkLightBlueWallet: '#0073E6',
-      bannerLinkDarkBlueWallet: '#0066CC',
-      blueIO: {
-        500: '#0B3EE3',
-      },
-      blueWallet: {
-        500: '#0066CC',
-      },
-      highLightBoxLightGreenBackground: '#EFFBFF',
-      highLightBoxPurpleText: '#5517E3',
-      highLightBoxLightPurpleBackground: '#DDD1F9',
-      black: '#000000',
-      white: '#FFF',
-      matteWhiteBorder: '#ffffff80',
-      ioChipsBackgroundColorLight: '#0B3EE3',
-      walletChipsBackgroundColorLight: '#0066CC',
-      sendChipsBackgroundColorLight: '#0073E6',
-      chipsBackgroundColorDark: '#ebebf52e',
-      ioChipsBackgroundColorLightHover: '#0A34CC',
-      walletChipsBackgroundColorLightHover: '#004D99',
-      sendChipsBackgroundColorLightHover: '#0066CC',
-      chipsBackgroundColorDarkHover: '#ebebf54d',
-      chipsTextColor: '#ffffff',
-      borderColor: '#E8EBF1',
-      outlineColor: '#0073e6',
-      walletOutlineColor: '#0066CC',
-      ctaContainedLightHoverColor: '#D9D9D9',
-      ctaContainedDarkHoverColor: '#0055AA',
-    },
-  },
-};
-
-const fontFamily = FONT_THEME_EXPERIMENTAL;
-
-export const themeExperimental = createTheme(MUIItaliaTheme, {
-  typography: {
-    fontFamily,
-    body1: { fontFamily },
-    body2: { fontFamily },
-    h1: { fontFamily },
-    h2: { fontFamily },
-    h3: { fontFamily },
-    h4: { fontFamily },
-    h5: { fontFamily },
-    h6: { fontFamily },
-    button: { fontFamily },
-    overline: { fontFamily },
-    subtitle1: { fontFamily },
-    subtitle2: { fontFamily },
-    caption: { fontFamily },
-  },
-  ...themeStyles,
-});
-
-export const themeBase = createTheme(MUIItaliaTheme, { ...themeStyles });
-
-export const theme = createTheme(MUIItaliaTheme, { ...themeStyles });
